move voucher publish to Distributions API
